server changes

+++ src/client/UpdateTrackerThread.java
@@ -49,9 +49,9 @@ public class UpdateTrackerThread extends Thread
 			Object[] myQueueArray = myQueue.toArray();
 			for(int lcv = 0; lcv < myQueueArray.length ; lcv++)
 			{
-				req.setStartBytes((int) myQueueArray[lcv]);
+				req.setStartBytes((Long) myQueueArray[lcv]);
 				lcv++;
-				req.setEndBytes((int) myQueueArray[lcv]);
+				req.setEndBytes((Long) myQueueArray[lcv]);
 				
 				out.print(req.toString());
 				String resp = "";
